+++ R/algorithm.R
@@ -155,6 +155,42 @@ algorithm <- function() {
       title = "Metformin purchases that aren't potentially for the treatment of PCOS",
       logic = "NOT (koen == 2 AND atc =~ '^A10BA02$' AND ((date - foed_dato) < years(40) OR indication_code %in% c('0000092', '0000276', '0000781')))",
       comments = "Woman is defined as 2 in `koen`."
+    ),
+    has_t1d = list(
+      register = NA,
+      title = "Classifying type 1 diabetes status",
+      logic = "(has_only_insulin_purchases & has_any_t1d_primary_diagnosis) | (!has_only_insulin_purchases & has_majority_t1d_diagnoses & has_two_thirds_insulin & has_insulin_purchases_within_180_days)",
+      comments = "The final classification for type 1 diabetes. Depends on all the previous steps to create these intermediate logical variables."
+    ),
+    has_any_t1d_primary_diagnosis = list(
+      register = NA,
+      title = "Any primary diagnosis for type 1 diabetes",
+      logic = "(n_t1d_endocrinology + n_t1d_medical) >= 1",
+      comments = "This is used to classify type 1 diabetes. Naturally, having any type 1 diabetes diagnosis is indicative of type 1 diabetes."
+    ),
+    has_majority_t1d_diagnoses = list(
+      register = NA,
+      title = "Determine if the majority of diagnoses are for type 1 diabetes",
+      logic = "if_else(n_t1d_endocrinology + n_t2d_endocrinology > 0, n_t1d_endocrinology > n_t2d_endocrinology, n_t1d_medical > n_t2d_medical)",
+      comments = "This is used to classify type 1 diabetes. Endocrinology diagnoses are prioritised if available, otherwise other medical department diagnoses are used. If no diabetes type-specific primary diagnoses are available from an endocrinology or other medical departments, this variable is returned as `FALSE`."
+    ),
+    has_two_thirds_insulin = list(
+      register = NA,
+      title = "Whether two-thirds of GLD doses are insulin doses",
+      logic = "(n_insulin_doses / n_gld_doses) >= 2/3",
+      comments = "This is used to classify type 1 diabetes. If multiple types of GLD are purchased, this indicates if at least two-thirds are insulin, which is important to determine type 1 diabetes status."
+    ),
+    has_only_insulin_purchases = list(
+      register = NA,
+      title = "Whether only insulin was purchased as a GLD",
+      logic = "n_insulin_doses >= 1 & n_insulin_doses == n_gld_doses",
+      comments = "This is used to classify type 1 diabetes. If only insulin is purchased, this is a strong reason to suspect type 1 diabetes."
+    ),
+    has_insulin_purchases_within_180_days = list(
+      register = NA,
+      title = "Whether any insulin was purchased within 180 days of the first purchase of GLD",
+      logic = "any(is_insulin_gld_code & date <= (first_gld_date + days(180)))",
+      comments = "This is used to classify type 1 diabetes. It determines if any insulin was bought shortly after first buying any type of GLD, which suggests type 1 diabetes."
     )
   )
 }
@@ -167,11 +203,6 @@ algorithm <- function() {
 #' @return A character string.
 #' @keywords internal
 #'
-#' @examples
-#' \dontrun{
-#' get_algorithm_logic("hba1c")
-#' get_algorithm_logic("gld")
-#' }
 get_algorithm_logic <- function(logic_name, algorithm = NULL) {
   checkmate::assert_character(logic_name)
   if (!is.null(algorithm)) {
@@ -189,4 +220,19 @@ get_algorithm_logic <- function(logic_name, algorithm = NULL) {
       "([a-zA-Z0-9_]+) \\=\\~ '(.*?)'",
       "stringr::str_detect(\\1, '\\2')"
     )
+}
+
+#' Parse the logic strings into R expressions
+#'
+#' @param logic The name of the logic to use.
+#'
+#' @return An R expression.
+#' @keywords internal
+#'
+logic_as_expression <- function(logic) {
+  logic |>
+    rlang::set_names() |>
+    purrr::map(get_algorithm_logic) |>
+    # To convert the string into an R expression
+    purrr::map(rlang::parse_expr)
 }