+++ tests/testthat/test-exclude-pregnancy.R
@@ -21,14 +21,15 @@ lpr3 <- prepare_lpr3(
 
 no_pcos <- register_data$lmdb |>
   include_gld_purchases() |>
+  add_insulin_purchases_cols() |>
   exclude_potential_pcos(register_data$bef)
 
-preg_dates <- get_pregnancy_dates(lpr2, lpr3)
+preg_dates <- keep_pregnancy_dates(lpr2, lpr3)
 
 hba1c <- include_hba1c(register_data$lab_forsker)
 
 test_that("pregnancy events are excluded as expected", {
-  actual <- exclude_pregnancy(
+  actual <- exclude_pregnancies(
     excluded_pcos = no_pcos,
     pregnancy_dates = preg_dates,
     included_hba1c = hba1c